[TeamC] refactor question row and refactor question and answer widgets into assessment_data_editor widget
